Got cube rendering, gotta fix UVs and use wrapper objects

+++ Headers/mesh.h
@@ -9,14 +9,14 @@ class Texture;
 class Mesh {
 private:
 	std::vector<size_t> mTextureIndices;
-	void setupMesh(const std::vector<Vertex>& vertices, const std::vector<size_t>& indices);
+	void setupMesh(const std::vector<Vertex>& vertices, const std::vector<unsigned int>& indices);
 
 public:
 	unsigned int mVAO, mVBO, mEBO;
 	unsigned int mVertexCount;
 	float mShininess;
 
-	Mesh(const std::vector<Vertex>& vertices, const std::vector<size_t>& indices, const std::vector<size_t>& textureIndices, float shininess);
+	Mesh(const std::vector<Vertex>& vertices, const std::vector<unsigned int>& indices, const std::vector<size_t>& textureIndices, float shininess);
 
 	int getFirstDiffuse(const std::vector<Texture>& textures) const;
 	int getFirstSpecular(const std::vector<Texture>& textures) const;
